requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==4.2.20
gunicorn==23.0.0
indic-transliteration==2.3.69
lxml==5.3.1
markdown-it-py==3.0.0
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sqlparse==0.5.3
toml==0.10.2
typer==0.15.2
typing-extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
